requirements.txt
PyYAML
Faker